requirements/development.txt
-r base.txt

coverage==5.2.1
django-coverage-plugin==1.8.0
django-debug-toolbar==2.2
hypothesis==5.30.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-mock==3.3.1
pytest-sugar==0.9.4
pytest==5.4.3
